WIP: POI and Grid utils out of old files #1
